#19 Allow scrollspy inside of scroll context
This is a somewhat naive implementation, and requires the scrollable
container to be positioned with `relative` or `absolute` (because we're
using `offsetTop` and `offsetHeight` and `offsetParent` to determine the
"spied" element's position within its container.

A more robust solution can/should be explored.  However, the current
implementation appears to work, and the context.html demo has been
updated to include spies.

Feedback is always welcome!

#19