Fix falsy error.
When doing `element.scrollTop(number, 0)`, due to an incorrect falsy value check, 0 would not be considered a valid animation duration. Suppose one wants to animate for 0 seconds, or at the very least expects a promise to be consistently returned from the function. This fixes that by allowing a duration of 0 seconds.